third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/rootscroller-origin-trial-interfaces-script-added.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>RootScroller - origin trial is enabled by script-added meta tag</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/origin-trials-helper.js"></script>
<script>
// Generate token with the command:
// generate_token.py  http://127.0.0.1:8000 RootScroller --expire-timestamp=2000000000
var token = "AuKypT9HyT7fv0cU54WMd1bUmdBYTFED3hKRFWfvxauV8saEM1CoOFLXDd3ApWwJCvimlrGVcqj+0Z9rJL8uFgIAAABUeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiUm9vdFNjcm9sbGVyIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9";

let properties_to_check = {'Document': ['rootScroller']};

test(t => {
  OriginTrialsHelper.check_properties_missing(this, properties_to_check);
}, "RootScroller related properties on interfaces before adding trial token via script.");

OriginTrialsHelper.add_token(token);

test(t => {
  OriginTrialsHelper.check_properties(this, properties_to_check);
}, "RootScroller related properties on interfaces after adding trial token via script.");
</script>
