third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/rootscroller-origin-trial-interfaces-script-added-expected.txt

This is a testharness.js-based test.
FAIL RootScroller related properties on interfaces before adding trial token via script. assert_equals: Property rootScroller exists on Document expected false but got true
PASS RootScroller related properties on interfaces after adding trial token via script. 
Harness: the test ran to completion.
